Add files via upload
Write a recursive C program to implement the Tower of Hanoi game. The Poles are labelled as 0, 1 and 2 and the disks are labelled from 1 to n
in the increasing order of their diameters.Input has a single line with the number of disks. The i'th line of the output shows the i'th movement made in the game.