Add files via upload
C program that takes as input an integer n and an integer array A of length n, and sorts the array in the non-decreasing order using Insertion Sort algorithm. Make sure your program has two functions other than the main function. One function which performs insertion sort and the other for insertion which takes the starting address of an array A and an index i as parameters (assuming A[0,⋯,i−1] is in non-decreasing order) and inserts A[i] into its appropriate position so that A[0,⋯,i] are now in non-decreasing order of the corresponding elements in A, while elements of A after index i are unaffected.