Add files via upload
A) Write a function that takes an integer n and an array A of size n as arguments, to find the index of the largest element from A
and exchange the element in that position with the element in the last position. 
The selection sort algorithm is a well known algorithm for sorting. It works by repeatedly calling the function in the step (A). After the first iteration, the largest element will be in the last position. After the second iteration, the second largest element will be in the second last position.  After n−1
iteration, the input array becomes sorted.
         B) Write a C program to implement selection sort using this function.